requirements.txt
attrs
spacy==3.5.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
numpy
nltk
gensim
pandas
pyinflect